Fix ci

+++ src/Controllers/Payloads/Metric.php
@@ -6,13 +6,13 @@ final class Metric
 {
     public string $title;
 
-    public float $value;
+    public int $value;
 
     public ?int $previousValue;
 
     public string $format;
 
-    public function __construct(string $title, float $value = 0, ?int $previousValue = null, string $format = '%d')
+    public function __construct(string $title, int $value = 0, ?int $previousValue = null, string $format = '%d')
     {
         $this->title = $title;
         $this->value = $value;
